Added sproutseo template on settings layout which display a form text macro for inputting csv input to be converted to json.
Added generateController that process the conversion of csv text to json format.

Added empty and invalid csv input validation and displays error notice to the setting template.

Added json format text after process the csv text on the sprout seo integration page. 

Created script that removes header data if found and not remove first line of row if it is not found. #1233379