Created process weed controller method that removes seeded and imported data.
Updated element importers on how they returun element ids to svoid error on seeding.

Created getSeedsByIds method that gets seeds by ids which is submitted from the weed row.

Created script that removes the data which is seed or imported and also allow them to keep the data.

Added script that weeds or removes all and keeps all seeded and imported data.